requirements.txt
flask
gunicorn
google-genai>=0.8.0
python-dotenv==1.0.0
redis